changed the order of reading and writing message
if you want to make the chat perfect, crate a copy of this message.php and rename it something like message_read.php. Cut line 21-32 from message.php and paste in message_read.php. and edit message.html to load message_read.php only when message should be read and not written to database.